basic prototype of executor
