+++ src/executor/mod.rs
@@ -1,4 +1,5 @@
 mod executor;
+pub mod storage;
 pub mod task;
 
 pub use executor::run;
