+++ tests/executor_tests.rs
@@ -1,7 +1,8 @@
 #[cfg(test)]
 mod tests {
     use async_trait::async_trait;
-    use capp::executor::task::{Task, TaskProcessor, TaskStorage};
+    use capp::executor::storage::TaskStorage;
+    use capp::executor::task::{Task, TaskProcessor};
     use capp::executor::{self, ExecutorOptions};
     use serde::{Deserialize, Serialize};
     use std::collections::{HashMap, VecDeque};
@@ -75,7 +76,7 @@ mod tests {
     impl TaskStorage<TaskData, TaskError> for TestStorage {
         async fn hashmap_set(
             &self,
-            task: &Task<TaskData, TaskError>,
+            task: &Task<TaskData>,
         ) -> Result<(), TaskError> {
             let mut hashmap = self.hashmap.lock().unwrap();
             let task_data = serde_json::to_string(task).unwrap();
@@ -86,18 +87,14 @@ mod tests {
         async fn hashmap_get(
             &self,
             task_id: Uuid,
-        ) -> Result<Option<Task<TaskData, TaskError>>, TaskError> {
+        ) -> Result<Option<Task<TaskData>>, TaskError> {
             let hashmap = self.hashmap.lock().unwrap();
             let data = hashmap.get(&task_id).unwrap();
-            let task_data: Task<TaskData, TaskError> =
-                serde_json::from_str(data).unwrap();
+            let task_data: Task<TaskData> = serde_json::from_str(data).unwrap();
             Ok(Some(task_data))
         }
 
-        async fn list_push(
-            &self,
-            task: &Task<TaskData, TaskError>,
-        ) -> Result<(), TaskError> {
+        async fn list_push(&self, task: &Task<TaskData>) -> Result<(), TaskError> {
             let mut list = self.list.lock().unwrap();
             let mut hashmap = self.hashmap.lock().unwrap();
 
@@ -107,16 +104,13 @@ mod tests {
             Ok(())
         }
 
-        async fn list_pop(
-            &self,
-        ) -> Result<Option<Task<TaskData, TaskError>>, TaskError> {
+        async fn list_pop(&self) -> Result<Option<Task<TaskData>>, TaskError> {
             let mut list = self.list.lock().unwrap();
             let hashmap = self.hashmap.lock().unwrap();
 
             if let Some(task_id) = list.pop_front() {
                 let task_data = hashmap.get(&task_id).unwrap();
-                let task: Task<TaskData, TaskError> =
-                    serde_json::from_str(task_data).unwrap();
+                let task: Task<TaskData> = serde_json::from_str(task_data).unwrap();
                 return Ok(Some(task));
             }
             Ok(None)
@@ -124,8 +118,8 @@ mod tests {
 
         async fn ack(
             &self,
-            task: &Task<TaskData, TaskError>,
-        ) -> Result<Task<TaskData, TaskError>, TaskError> {
+            task: &Task<TaskData>,
+        ) -> Result<Task<TaskData>, TaskError> {
             let mut hashmap = self.hashmap.lock().unwrap();
             let task_data = hashmap.remove(&task.task_id).unwrap();
             let task = serde_json::from_str(&task_data).unwrap();
@@ -143,7 +137,7 @@ mod tests {
         let rt = Runtime::new().unwrap();
 
         for i in 1..=3 {
-            let task: Task<TaskData, TaskError> = Task::new(TaskData {
+            let task: Task<TaskData> = Task::new(TaskData {
                 domain: "one".to_string(),
                 value: i,
                 finished: false,
@@ -152,7 +146,7 @@ mod tests {
         }
 
         for i in 1..=3 {
-            let task: Task<TaskData, TaskError> = Task::new(TaskData {
+            let task: Task<TaskData> = Task::new(TaskData {
                 domain: "two".to_string(),
                 value: i * 3,
                 finished: false,
@@ -161,7 +155,7 @@ mod tests {
         }
 
         for _ in 1..=3 {
-            let task: Task<TaskData, TaskError> = Task::new(TaskData {
+            let task: Task<TaskData> = Task::new(TaskData {
                 domain: "three".to_string(),
                 value: 2,
                 finished: false,
